Added new functionality, revised the code

+++ __main__.py
@@ -2,14 +2,15 @@ import shutil
 import errno
 import argparse
 from filecmp import dircmp
-from pathlib import Path, PurePath
+from pathlib import Path
 from json import JSONDecoder
 from collections import OrderedDict
+from abc import ABC, abstractmethod
 
 decoder = JSONDecoder(object_pairs_hook=OrderedDict)
 
 
-class Task:
+class Task(ABC):
 
     def __init__(self, source, target):
         self.source, self.target = source, target
@@ -22,21 +23,43 @@ class Task:
 
     def diff(self):
         cmp = dircmp(str(self.source), str(self.target))
-        yield from self._recursive_cmp(PurePath(), cmp)
+        yield from self._recursive_cmp(Path(), cmp)
 
-    def copy(self, path):
-        src, dest = str(self.source / path), str(self.target / path)
-        try:
-            shutil.copy(src, dest)
-        except OSError as e:
-            if e.errno == errno.EISDIR:
-                shutil.copytree(src, dest)
-            else:
-                raise e from None
+    @abstractmethod
+    def run(confirm=False):
+        pass
 
-    def copy_all(self):
+
+class CopyTask(Task):
+
+    def run(self, confirm=False):
         for path in self.diff():
-            self.copy(path)
+            src, dest = str(self.source / path), str(self.target / path)
+            if confirm and input('Copy {} to {}? '.format(src, dest)) not in 'yY':
+                continue
+            try:
+                shutil.copy(src, dest)
+            except OSError as e:
+                if e.errno == errno.EISDIR:
+                    shutil.copytree(src, dest)
+                else:
+                    raise e from None
+
+
+class RemovalTask(Task):
+
+    def run(self, confirm=False):
+        for path in self.diff():
+            path = self.source / path
+            if confirm and input('Remove {}?'.format(path)) not in 'yY':
+                continue
+            try:
+                path.unlink()
+            except OSError as e:
+                if e.errno == errno.EISDIR:
+                    shutil.rmtree(str(path))
+                else:
+                    raise e from None
 
 
 parser = argparse.ArgumentParser()
@@ -44,30 +67,34 @@ parser.add_argument('source')
 parser.add_argument('targets', nargs='*')
 parser.add_argument('--confirm', '-c', action='store_true')
 parser.add_argument('--list', '-l',  action='store_true')
+parser.add_argument('--remove-extra', '-r', action='store_true')
 args = parser.parse_args()
 
 source = Path(args.source).resolve()
 tasks = OrderedDict()
 
+if args.list and (args.confirm or args.remove_extra):
+    raise RuntimeError('--list/-l must be sole argument')
+
+# Can be a path to the source directory or
+# a path to the configutation file
 if not source.is_dir():
     config = decoder.decode(source.read_text())
     for k, v in config.items():
-        tasks[k] = [Task(v['source'], Path(t).resolve()) for t in v['targets']]
+        tasks[k] = [CopyTask(v['source'], Path(t).resolve()) for t in v['targets']]
 else:
-    tasks[source.stem] = [Task(source, Path(t).resolve()) for t in args.targets]
+    tasks[source.stem] = [CopyTask(source, Path(t).resolve()) for t in args.targets]
 
-for k, v in tasks.items():
+if args.remove_extra:
+    for k, v in tasks.items():
+        v.extend([RemovalTask(t.target, t.source) for t in v])
+
+for _, v in tasks.items():
     for task in v:
         if args.list:
             diff = list(task.diff())
             lines = '\n'.join(['  * {}'.format(path) for path in diff])
             print('Difference between {} and {}'.format(task.source, task.target),
                   ' is empty.' if not diff else ':\n' + lines, sep='')
-        elif args.confirm:
-            for path in task.diff():
-                src, dest = task.source / path, (task.target / path).parent
-                if input('Copy {} to {}? '.format(src, dest)) not in 'yY':
-                    continue
-                task.copy(path)
         else:
-            task.copy_all()
+            task.run(args.confirm)
